Update: Refactor add post form
