+++ src/shared/ui/components/post-form/index.ts
@@ -1,0 +1,1 @@
+export { PostForm } from './post-form';
