+++ src/widgets/add-post-form/api/fetch-labels.ts
