+++ src/widgets/add-post-form/api/schema.ts
