+++ src/widgets/add-post-form/index.ts
@@ -1,0 +1,1 @@
+export { PostFormContainer } from './ui/add-post-form-container';
