--- src/widgets/post-form/index.ts
@@ -1,1 +1,0 @@
-export { PostFormContainer } from './ui/post-form-container';
